requirements.txt
absl-py==2.2.2
albucore==0.0.24
albumentations==2.0.7
annotated-types==0.7.0
attrs==25.3.0
blinker==1.9.0
cachetools==5.5.2
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
colorama==0.4.6
contourpy==1.3.2
cycler==0.12.1
defusedxml==0.7.1
filelock==3.13.1
flask==3.1.1
flatbuffers==25.2.10
fonttools==4.58.0
fpdf2==2.8.3
fsspec==2024.6.1
google-ai-generativelanguage==0.6.15
google-api-core==2.25.0
google-api-python-client==2.171.0
google-auth==2.40.3
google-auth-httplib2==0.2.0
google-generativeai==0.8.5
googleapis-common-protos==1.70.0
grpcio==1.72.1
grpcio-status==1.71.0
httplib2==0.22.0
idna==3.10
itsdangerous==2.2.0
jax==0.6.1
jaxlib==0.6.1
jinja2==3.1.6
kiwisolver==1.4.8
markupsafe==3.0.2
matplotlib==3.10.3
ml-dtypes==0.5.1
mpmath==1.3.0
networkx==3.3
numpy==1.26.4
opencv-contrib-python==4.11.0.86
opencv-python==4.11.0.86
opencv-python-headless==4.11.0.86
opt-einsum==3.4.0
packaging==25.0
pillow==11.0.0
proto-plus==1.26.1
protobuf==5.29.5
pyasn1==0.6.1
pyasn1-modules==0.4.2
pycparser==2.22
pydantic==2.11.5
pydantic-core==2.33.2
pyparsing==3.2.3
python-dateutil==2.9.0.post0
pyyaml==6.0.2
requests==2.32.3
rsa==4.9.1
scipy==1.15.3
sentencepiece==0.2.0
setuptools==70.2.0
simsimd==6.2.1
six==1.17.0
sounddevice==0.5.2
stringzilla==3.12.5
sympy==1.13.1
torch==2.5.1+cu121
torchaudio==2.5.1+cu121
torchvision==0.20.1+cu121
tqdm==4.67.1
typing-extensions==4.12.2
typing-inspection==0.4.1
uritemplate==4.2.0
urllib3==2.4.0
werkzeug==3.1.3
